Cleanup and ESP modifications
